+++ src/components/Store.vue
@@ -154,6 +154,7 @@ import { highlight, languages } from 'prismjs/components/prism-core'
 import 'prismjs/components/prism-clike'
 import 'prismjs/components/prism-javascript'
 import 'prismjs/themes/prism-tomorrow.css'
+import { mapMutations } from 'vuex'
 
 export default {
   name: 'Store',
@@ -185,9 +186,7 @@ export default {
     }
   },
   methods: {
-    showSnackbar (text) {
-      this.$emit('showSnackbar', text)
-    },
+    ...mapMutations(['showSnackbar']),
     async deleteFile (item) {
       if (
         await this.$listeners.showConfirm(
